// src/store/modules/jt-cesium-vue/modules/toolbar/mutation-types.ts
export enum ToolbarMutationTypes {
  RESET_STATE = 'RESET_STATE',

  SET_VISIBLE = 'SET_VISIBLE',
  SET_DROP_DOWN = 'SET_DROP_DOWN',

  // nature
  SET_SHOW_SUN = 'SET_SHOW_SUN',
  SET_SHOW_MOON = 'SET_SHOW_MOON',
  SET_SHOW_SKY_ATMOSPHERE = 'SET_SHOW_SKY_ATMOSPHERE',
  SET_ENABLE_LIGHT = 'SET_ENABLE_LIGHT',
  SET_SHOW_SKY_BOX = 'SET_SHOW_SKY_BOX',
  SET_SHOW_SHADOW = 'SET_SHOW_SHADOW',

  // other
  SET_DEPTH_TEST_AGAINST_TERRAIN = 'SET_DEPTH_TEST_AGAINST_TERRAIN',

  // draw
  SET_DRAW_POINT_ACTIVE = 'SET_DRAW_POINT_ACTIVE',
  SET_DRAW_POLYLINE_ACTIVE = 'SET_DRAW_POLYLINE_ACTIVE',
  SET_DRAW_POLYGON_ACTIVE = 'SET_DRAW_POLYGON_ACTIVE',

  // measure
  SET_MEASURE_POINT_ACTIVE = 'SET_MEASURE_POINT_ACTIVE',
  SET_MEASURE_POLYLINE_ACTIVE = 'SET_MEASURE_POLYLINE_ACTIVE',
  SET_MEASURE_POLYGON_ACTIVE = 'SET_MEASURE_POLYGON_ACTIVE',

  // tool3dtile
  SET_HIGHLIGHT_3DTILE_FEATURE_ACTIVE = 'SET_HIGHLIGHT_3DTILE_FEATURE_ACTIVE',
  SET_HOVER_CLASSIFICATION_ACTIVE = 'SET_HOVER_CLASSIFICATION_ACTIVE',
  SET_CLICK_CLASSIFICATION_ACTIVE = 'SET_CLICK_CLASSIFICATION_ACTIVE',

  // elevation contour
  SET_ELEVATION_CONTURE_ACTIVE = 'SET_ELEVATION_CONTURE_ACTIVE',

  // terrain sampling
  SET_TERRAIN_SAMPLING = 'SET_TERRAIN_SAMPLING',
}