// src/store/modules/jt-cesium-vue/modules/toolbar/action-types.ts
export enum ToolbarActionTypes {
  RESET_STATE = 'RESET_STATE',

  SET_VISIBLE = 'SET_VISIBLE',
  SET_DROP_DOWN = 'SET_DROP_DOWN',

  // nature
  SWITCH_SHOW_SUN = 'SWITCH_SHOW_SUN',
  SWITCH_SHOW_MOON = 'SWITCH_SHOW_MOON',
  SWITCH_SHOW_SKY_ATMOSPHERE = 'SWITCH_SHOW_SKY_ATMOSPHERE',
  SWITCH_ENABLE_LIGHT = 'SWITCH_ENABLE_LIGHT',
  SWITCH_SHOW_SKY_BOX = 'SWITCH_SHOW_SKY_BOX',

  // other
  SWITCH_DEPTH_TEST_AGAINST_TERRAIN = 'SWITCH_DEPTH_TEST_AGAINST_TERRAIN',

  // draw
  SET_DRAW_POINT_ACTIVE = 'SET_DRAW_POINT_ACTIVE',
  SET_DRAW_POLYLINE_ACTIVE = 'SET_DRAW_POLYLINE_ACTIVE',
  SET_DRAW_POLYGON_ACTIVE = 'SET_DRAW_POLYGON_ACTIVE',

  // measure
  SET_MEASURE_POINT_ACTIVE = 'SET_MEASURE_POINT_ACTIVE',
  SET_MEASURE_POLYLINE_ACTIVE = 'SET_MEASURE_POLYLINE_ACTIVE',
  SET_MEASURE_POLYGON_ACTIVE = 'SET_MEASURE_POLYGON_ACTIVE',

  // tool3dtile
  SET_HIGHLIGHT_3DTILE_FEATURE_ACTIVE = 'SET_HIGHLIGHT_3DTILE_FEATURE_ACTIVE',
  SET_HOVER_CLASSIFICATION_ACTIVE = 'SET_HOVER_CLASSIFICATION_ACTIVE',
  SET_CLICK_CLASSIFICATION_ACTIVE = 'SET_CLICK_CLASSIFICATION_ACTIVE',

  // elevation contour
  SET_ELEVATION_CONTURE_ACTIVE = 'SET_ELEVATION_CONTURE_ACTIVE',

  // terrain sampling
  SET_TERRAIN_SAMPLING = 'SET_TERRAIN_SAMPLING',
}